app/src/main/python/requirements-android.txt
Flask==2.3.3
Werkzeug==2.3.7
gspread==5.10.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
bcrypt==4.0.1
python-dotenv==1.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
